pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = ""
authors = ["Irlen Turlykhanov <irlen.turlykhanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.3"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.2.3"
langchain-openai = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jarvis = "jarvis.__main__:main"